+++ app/config/auth.config.js
@@ -2,5 +2,5 @@ require('dotenv').config();
 
 module.exports = {
   // Generated here : https://www.grc.com/passwords.htm
-  SECRET: process.env.JWT_SECRET,
+  secret: process.env.JWT_SECRET,
 };
