--- app/config/server.config.js
@@ -1,6 +1,0 @@
-require('dotenv').config();
-
-module.exports = {
-  PRODUCTION: process.env.PRODUCTION === 'true',
-  PORT: 8080,
-};
