+++ app/controllers/event.controller.js
@@ -1,3 +1,4 @@
+const config = require('../config/auth.config');
 const db = require('../models/db.model');
 const {
   unexpectedErrorCatch,
