+++ app/middlewares/user.middleware.js
@@ -16,9 +16,8 @@ const verifyAccessToken = (req, res, next) => {
     });
   }
 
-  jwt.verify(token, config.SECRET, (err, decoded) => {
+  jwt.verify(token, config.secret, (err, decoded) => {
     if (err) {
-      console.log(err);
       return res.status(401).json({
         message: 'Unauthorized!',
       });
